+++ src/lib/webview/controller/bbIssue/createBitbucketIssueWebviewController.ts
@@ -33,6 +33,8 @@ export class CreateBitbucketIssueWebviewController implements WebviewController<
         private analytics: AnalyticsApi,
     ) {}
 
+    public onShown(): void {}
+
     public title(): string {
         return 'Create Bitbucket Issue';
     }