+++ src/lib/webview/controller/onboarding/onboardingWebviewController.ts
@@ -1,14 +1,17 @@
 import { defaultActionGuard } from '@atlassianlabs/guipi-core-controller';
-import { Experiments } from 'src/util/featureFlags';
+// eslint-disable-next-line no-restricted-imports
+import { ConfigurationTarget, WebviewPanel } from 'vscode';
 
-import { isBasicAuthInfo } from '../../../../atlclients/authInfo';
+import { configuration } from '../../../../../src/config/configuration';
+// eslint-disable-next-line no-restricted-imports
+import { Container } from '../../../../../src/container';
+import { isBasicAuthInfo, ProductBitbucket, ProductJira } from '../../../../atlclients/authInfo';
 import { AnalyticsApi } from '../../../analyticsApi';
 import { CommonAction, CommonActionType } from '../../../ipc/fromUI/common';
 import { OnboardingAction, OnboardingActionType } from '../../../ipc/fromUI/onboarding';
 import { WebViewID } from '../../../ipc/models/common';
 import { CommonMessage, CommonMessageType } from '../../../ipc/toUI/common';
-import { SectionChangeMessage } from '../../../ipc/toUI/config';
-import { OnboardingMessage, OnboardingMessageType } from '../../../ipc/toUI/onboarding';
+import { OnboardingInitMessage, OnboardingMessage, OnboardingMessageType } from '../../../ipc/toUI/onboarding';
 import { Logger } from '../../../logger';
 import { formatError } from '../../formatError';
 import { CommonActionMessageHandler } from '../common/commonActionMessageHandler';
@@ -18,9 +21,9 @@ import { OnboardingActionApi } from './onboardingActionApi';
 export const id: string = 'atlascodeOnboardingV2';
 const title: string = 'Getting Started';
 
-export class OnboardingWebviewController implements WebviewController<SectionChangeMessage> {
+export class OnboardingWebviewController implements WebviewController<OnboardingInitMessage> {
     public readonly requiredFeatureFlags = [];
-    public readonly requiredExperiments = [Experiments.NewAuthUI];
+    public readonly requiredExperiments = [];
 
     private _messagePoster: MessagePoster;
     private _api: OnboardingActionApi;
@@ -49,6 +52,18 @@ export class OnboardingWebviewController implements WebviewController<SectionCha
         this._messagePoster(message);
     }
 
+    public async onShown(panel: WebviewPanel): Promise<void> {
+        try {
+            await configuration.update('jira.enabled', undefined, ConfigurationTarget.Global);
+        } catch {}
+
+        // focus the atlassian extension panels when the onboarding view shows...
+        Container.focus();
+
+        // ...then focus back here
+        panel.reveal(undefined, false);
+    }
+
     public title(): string {
         return title;
     }
@@ -58,30 +73,27 @@ export class OnboardingWebviewController implements WebviewController<SectionCha
     }
 
     public async onSitesChanged(): Promise<void> {
-        const [jiraSites, bbSites] = await this._api.getSitesWithAuth();
         this.postMessage({
             type: OnboardingMessageType.SitesUpdate,
-            jiraSites: jiraSites,
-            bitbucketSites: bbSites,
+            jiraSitesConfigured: Container.siteManager.productHasAtLeastOneSite(ProductJira),
+            bitbucketSitesConfigured: Container.siteManager.productHasAtLeastOneSite(ProductBitbucket),
         });
     }
 
     private async invalidate() {
-        const [jiraSites, bbSites] = await this._api.getSitesWithAuth();
         const target = this._api.getConfigTarget();
         const cfg = this._api.flattenedConfigForTarget(target);
         this.postMessage({
             type: OnboardingMessageType.Init,
-            bitbucketSites: bbSites,
-            jiraSites: jiraSites,
-            isRemote: this._api.getIsRemote(),
+            jiraSitesConfigured: Container.siteManager.productHasAtLeastOneSite(ProductJira),
+            bitbucketSitesConfigured: Container.siteManager.productHasAtLeastOneSite(ProductBitbucket),
             target: target,
             config: cfg,
         });
     }
 
-    public update() {
-        //No initial data to send!
+    public update(data: OnboardingInitMessage) {
+        this.invalidate();
     }
 
     public async onMessageReceived(msg: OnboardingAction | CommonAction) {