+++ src/lib/webview/controller/pullrequest/createPullRequestWebviewController.ts
@@ -38,6 +38,8 @@ export class CreatePullRequestWebviewController implements WebviewController<Wor
         this.initData = factoryData!;
     }
 
+    public onShown(): void {}
+
     public title(): string {
         return 'Create pull request';
     }