+++ src/lib/webview/controller/webviewController.ts
@@ -1,4 +1,6 @@
 import { Experiments, Features } from 'src/util/featureFlags';
+// eslint-disable-next-line no-restricted-imports
+import { WebviewPanel } from 'vscode';
 
 import { DetailedSiteInfo, Product } from '../../../atlclients/authInfo';
 
@@ -8,6 +10,7 @@ export interface WebviewController<FD> {
     requiredFeatureFlags: Features[];
     requiredExperiments: Experiments[];
 
+    onShown(panel: WebviewPanel): void;
     title(): string;
     screenDetails(): { id: string; site?: DetailedSiteInfo; product?: Product };
     onMessageReceived(msg: any): void;